Add files via upload
In this work, we discuss a variant of prize collecting travelling salesman’s problem considering the sets of salesmans, facilities and customers. Each facility can cover the patients that are within an exogenously defined distance known as coverage distance. Also, there is a limit on maximum distance travelled by each salesman while visiting the facilities. Further, a facility can be visited by at the most one salesman. Given this problem setting, the objective is to maximize the total number of customers covered by constructing Hamiltonian cycles of m salesmans over a subset of facilities. This problem is referred as the distance constrained maximal covering multi travelling salesman problem (DCMC-mTSP). This problem finds the potential applications in the areas such as humanitarian logistics and mobile healthcare applications in remote areas. Like the travelling salesman’s problem, the DCMC-mTSP is also NP-hard. Therefore, we propose a metaheuristic: a novel variant of ant-colony-optimization algorithm to efficiently solve the medium and the large size problem instances.